MNT: move Msg to utils
Still import at top level
