+++ bluesky/run_engine.py
@@ -20,8 +20,8 @@ from .utils import (CallbackRegistry, SignalHandler, normalize_subs_input,
                     AsyncInput, new_uid, sanitize_np, NoReplayAllowed,
                     RequestAbort, RequestStop,  RunEngineInterrupted,
                     IllegalMessageSequence, FailedPause, FailedStatus,
-                    InvalidCommand, PlanHalt)
-from . import Msg
+                    InvalidCommand, PlanHalt, Msg)
+
 from .plan_tools import ensure_generator
 from .plans import single_gen
